Support for empty cohortInclusion  table and BIGINT
https://github.com/OHDSI/CohortDiagnostics/commit/c482f28001fa6be89026f5ec228d3cf333cc956b#r42770785

+++ inst/sql/sql_server/inclusionStatsTables.sql
@@ -8,14 +8,14 @@ IF OBJECT_ID('tempdb..#cohort_summary_stats', 'U') IS NOT NULL
   DROP TABLE #cohort_summary_stats;
 
 CREATE TABLE #cohort_inc_result (
-	cohort_definition_id INT NOT NULL,
+	cohort_definition_id BIGINT NOT NULL,
 	inclusion_rule_mask BIGINT NOT NULL,
 	person_count BIGINT NOT NULL,
 	mode_id INT
 	);
 
 CREATE TABLE #cohort_inc_stats (
-	cohort_definition_id INT NOT NULL,
+	cohort_definition_id BIGINT NOT NULL,
 	rule_sequence INT NOT NULL,
 	person_count BIGINT NOT NULL,
 	gain_count BIGINT NOT NULL,
@@ -24,7 +24,7 @@ CREATE TABLE #cohort_inc_stats (
 	);
 
 CREATE TABLE #cohort_summary_stats (
-	cohort_definition_id INT NOT NULL,
+	cohort_definition_id BIGINT NOT NULL,
 	base_count BIGINT NOT NULL,
 	final_count BIGINT NOT NULL,
 	mode_id INT
